feat: estruturando o projeto
