+++ TechChallengeFIAP.Data/Repositorios/PessoaRepositorio.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using TechChallengeFIAP.Domain.Interfaces;
+
+namespace TechChallengeFIAP.Data.Repositorios;
+
+public class PessoaRepositorio : IPessoaRepositorio
+{
+
+}
