+++ TechChallengeFIAP.Data/TechChallengeFIAP.Data.csproj
@@ -1,17 +1,12 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\TechChallengeFIAP.Data\TechChallengeFIAP.Data.csproj" />
     <ProjectReference Include="..\TechChallengeFIAP.Domain\TechChallengeFIAP.Domain.csproj" />
   </ItemGroup>
