+++ TechChallengeFIAP.Domain/Entidades/Administrador.cs
@@ -1,0 +1,6 @@
+﻿namespace TechChallengeFIAP.Domain.Entidades;
+
+public class Administrador : Pessoas
+{
+
+}
