+++ TechChallengeFIAP.Domain/Entidades/Jogos.cs
@@ -1,0 +1,12 @@
+﻿namespace TechChallengeFIAP.Domain.Entidades;
+
+public class Jogos
+{
+    string NomeJogo { get; set; }
+    string DescricaoJogo { get; set; }
+    string ClassificacaoJogo { get; set; }
+    decimal ValorJogo { get; set; }
+    decimal ValorPromocao { get; set; }
+    DateTime DataLancamento { get; set; }
+    bool PromocaoHabilitada { get; set; }
+}
