+++ TechChallengeFIAP.Domain/Entidades/UsuarioComum.cs
@@ -1,0 +1,6 @@
+﻿namespace TechChallengeFIAP.Domain.Entidades;
+
+public class UsuarioComum : Pessoas
+{
+
+}
