+++ TechChallengeFIAP.Domain/Enums/Permissoes.cs
@@ -1,0 +1,7 @@
+﻿namespace TechChallengeFIAP.Domain.Enums;
+
+public enum Permissoes
+{
+    UsuarioComum = 0,
+    Administrador = 1
+}
