+++ TechChallengeFIAP.Domain/Interfaces/IPessoaRepositorio.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TechChallengeFIAP.Domain.Interfaces;
+
+public interface IPessoaRepositorio
+{
+
+}
