+++ TechChallengeFIAP.Domain/Interfaces/IUsuarios.cs
@@ -1,0 +1,12 @@
+﻿namespace TechChallengeFIAP.Domain.Interfaces;
+
+public interface IUsuarios
+{
+
+    string NomeCompleto { get; set; }
+    string NomeUsuario { get; set; }
+    string EmailUsuario { get; set; }
+    DateTime DataNascimento { get; set; }
+    string SenhaUsuario { get; set; }
+
+}
