+++ TechChallengeFIAP.sln
@@ -10,6 +10,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "0 - SolutionItems", "0 - So
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TechChallengeFIAP.Domain", "TechChallengeFIAP.Domain\TechChallengeFIAP.Domain.csproj", "{E1FC0DF5-A188-4A2C-8F48-57F5AB0B1126}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TechChallengeFIAP.Data", "TechChallengeFIAP.Data\TechChallengeFIAP.Data.csproj", "{E9BAE3AA-C371-4CBA-9095-23FF4C99B9C6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -20,6 +24,14 @@ Global
 		{B50EEE72-753F-45EE-B64F-4669FA03A126}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B50EEE72-753F-45EE-B64F-4669FA03A126}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B50EEE72-753F-45EE-B64F-4669FA03A126}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E1FC0DF5-A188-4A2C-8F48-57F5AB0B1126}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E1FC0DF5-A188-4A2C-8F48-57F5AB0B1126}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E1FC0DF5-A188-4A2C-8F48-57F5AB0B1126}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E1FC0DF5-A188-4A2C-8F48-57F5AB0B1126}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E9BAE3AA-C371-4CBA-9095-23FF4C99B9C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E9BAE3AA-C371-4CBA-9095-23FF4C99B9C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E9BAE3AA-C371-4CBA-9095-23FF4C99B9C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E9BAE3AA-C371-4CBA-9095-23FF4C99B9C6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
