+++ TechChallengeFIAP/GlobalUsings.cs
@@ -1,0 +1,3 @@
+﻿global using TechChallengeFIAP.Domain.Entidades;
+global using TechChallengeFIAP.Domain.Enums;
+global using TechChallengeFIAP.Domain.Interfaces;
