--- TechChallengeFIAP/Interface/IUsuarios.cs
@@ -1,14 +1,0 @@
-﻿namespace TechChallengeFIAP.Interface
-{
-    public interface IUsuarios
-    {
-
-        string NomeCompleto { get; set; }
-        string NomeUsuario { get; set; }
-        string EmailUsuario { get; set; }
-        DateTime DataNascimento { get; set; }
-        string SenhaUsuario { get; set; }
-
-    }
-
-}
