--- TechChallengeFIAP/Models/Administrador.cs
@@ -1,6 +1,0 @@
-﻿namespace TechChallengeFIAP.Models
-{
-    public class Administrador : Pessoas { 
-
-    }
-}
