--- TechChallengeFIAP/Models/Permissoes.cs
@@ -1,8 +1,0 @@
-﻿namespace TechChallengeFIAP.Models
-{
-    public enum Permissoes
-    {
-        UsuarioComum = 0,
-        Administrador = 1
-    }
-}
