--- TechChallengeFIAP/Models/UsuarioComum.cs
@@ -1,8 +1,0 @@
-﻿namespace TechChallengeFIAP.Models
-{
-    public class UsuarioComum : Pessoas
-    {
-
-    }
-
-}
